chore: Update search functionality in crypto widgets
